Rename mobility form (was .jsx.jsx)
